add new ui for user role

+++ inventaris/data.php
@@ -33,11 +33,18 @@ JOIN lokasi ON inventaris.id_lokasi = lokasi.id_lokasi
         <ul>
             <li><a href="../dashboard/index.php">Dashboard</a></li>
             <li><a href="data.php">Data Inventaris</a></li>
-            <li><a href="../lokasi/lokasi.php">Data Lokasi</a></li>
-            <li><a href="../peminjaman/index.php">Peminjaman</a></li>
-            <li><a href="../perbaikan/data_perbaikan.php">Perbaikan</a></li>
-            <li><a href="../laporan/laporan.php">Laporan</a></li>
-            <li><a href="../user/data_user.php">Manajemen User</a></li>
+
+            <?php if($isAdmin || $isTeknisi){ ?>
+                <li><a href="../lokasi/lokasi.php">Data Lokasi</a></li>
+                <li><a href="../peminjaman/index.php">Peminjaman</a></li>
+                <li><a href="../perbaikan/data_perbaikan.php">Perbaikan</a></li>
+                <li><a href="../laporan/laporan.php">Laporan</a></li>
+            <?php } ?>
+
+            <?php if($isAdmin){ ?>
+                <li><a href="../user/data_user.php">Manajemen User</a></li>
+            <?php } ?>
+
             <li><a href="../logout.php">Logout</a></li>
         </ul>
 
@@ -52,7 +59,10 @@ JOIN lokasi ON inventaris.id_lokasi = lokasi.id_lokasi
     
         <div class="table-container">
 
-            <a href="tambah.php" class="btn-tambah">+ Tambah Inventaris</a>
+            <!-- Tombol tambah hanya admin & teknisi -->
+            <?php if($isAdmin || $isTeknisi){ ?>
+                <a href="tambah.php" class="btn-tambah">+ Tambah Inventaris</a>
+            <?php } ?>
 
             <table>
     
@@ -87,8 +97,12 @@ JOIN lokasi ON inventaris.id_lokasi = lokasi.id_lokasi
                         <td><?php echo $row['nama_lokasi']; ?></td>
                         <td><?php echo $row['status']; ?></td>
                         <td>
-                            <a href="edit.php?id=<?php echo $row['id_barang']; ?>" class="btn-edit">Edit</a>
-                            <a href="hapus.php?id=<?php echo $row['id_barang']; ?>" class="btn-hapus"onclick="return konfirmasiHapus()">Hapus</a>
+
+                            <?php if($isAdmin || $isTeknisi){ ?>
+                                <a href="edit.php?id=<?php echo $row['id_barang']; ?>" class="btn-edit">Edit</a>
+                                <a href="hapus.php?id=<?php echo $row['id_barang']; ?>" class="btn-hapus" onclick="return konfirmasiHapus()">Hapus</a>
+                            <?php } ?>
+
                         </td>
 
                     </tr>
